Create a util file for "createAndReturnNewReferralId" function
Related to #120
